LayoutTests/platform/glib/accessibility/table-cell-display-block-expected.txt

This tests that if a table cell uses display:block, the table cell will still be accessible

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS cell1a.isEqual(cell1b) is true
PASS cell1a.role is 'AXRole: AXCell'
PASS domCell1.style.display is 'block'
PASS table2.role is 'AXRole: AXGroup'
FAIL cell2.role == 'AXRole: AXCell' should be false. Threw exception TypeError: null is not an object (evaluating 'cell2.role')
PASS successfullyParsed is true

TEST COMPLETE
